Wait for collect_initial to complete on k8s worker_management
The Kubernetes `MiqServer::WorkerManagement` class keeps a cache of pods
and deployments that is updated with two threads running
`watch_for_events`.  The way watches work on Kubernetes is you first run
a get on the collection which returns all current resources plus a
resourceVersion, you then pass this resourceVersion back to watch_entity
and kubernetes will send updates for the collection as of that
resourceVersion.

For cache consistency it is critical to know when the initial collection
is complete because before that you cannot trust that the cache has
finished being populated for the first time.

We can use a Concurrent::Event to wait for the thread to finish with the
initial collection, that way it is impossible for the worker_management
code to see a current_pods/deployments cache that isn't initially
consistent after calling `ensure_kube_monitors_started`.