Drop eth0 from MiqServer.get_network_information
Don't depend on having an interface named eth0 for the
get_network_information method.

MiqEnvironment already has logic for finding a non-loopback interface
with an ipv4 or ipv6 address.  This doesn't work great with multiple
interfaces but it does work on our default appliances with a single
non lo iface.