Add warning for local settings changes
The settings.local.yml file, while a powerful tool, has a downside in
that local changes modify all of dev/test/prod envs, without any
indication of change, and thus it's easily forgettable. This commit adds
a warning for the settings.local.yml file to indicate it has been
modified.