CP4AIOPS-3113 Introduce Settings.authentication.http_group_delimiter
This allows users to configure their LDAP group delimiter

Currently we are flexible and allow a , : or ;
But if a group has that value, then it causes issues.

The use case is to allow : in the group name for Liberty.
This allows us to specify a , as the delimiter for that case.