Unify Worker Environment Variables systemd/k8s
Handling of worker environment variables was spread across a number of
different locations and in a number of cases had to be duplicated for
systemd/kubernetes.

This adds a common `MiqWorker#environment_variables` method that applies
to all runtime platforms, as well as specific
systemd/container_environment_variables which are merged in depending on
the runtime environment.