Remove noisy test output logging we've never cared about
Technically, we could try to fix this and get the translators to update their data to have
this fixed, but it's possible there are situations where this happens again and it's not
an actual problem for gettext.

Even the original PR that added this said this overall number of newlines check
could flag intentionally changed values.  Since we've never reacted to this output,
we can remove it.

From that PR (we care about the 3rd check below):

1. if original string starts with a newline, translation needs to start with a newline too
2. if original string ends with a newline, translated string also needs to end with a newline
3. overal amount of newlines in original and translated string needs to match

In this PR, I am introducing a new test, which tests the above rules. The exception is the
3rd rule, since it seems that in some places the newline might have been added (or omitted)
intentionaly. So we're just logging these.

Reference:: https://github.com/ManageIQ/manageiq/pull/20815