simplify fix_auth tests
There is an issue in the way that encryption keys work.
You can sometimes decrypt a string that has been encoded with a different passkey.
You will get the wrong value out, but the system thinks that it succeeded.

Even though the odds of this are very low, since we run managiq so many times, we
run across it as a sporadic test failure.

This is an issue and we're removing the offending tests.
This functionality is tested in the `manageiq-password` gem.