Use YAML as default column serializer in rails 7.1
Originally, in https://www.github.com/rails/rails/pull/47422,
they wanted to remove the default of YAML for all applications.
It would be nil.

That was scaled back to only do this for new applications. Namely,
ones that call load_defaults 7.1.  It's supposed to require a coder
specified if you don't have a default.

The one that landed is this one: https://www.github.com/rails/rails/pull/47463