Poll more frequently in Ansible::Runner::ResponseAsync#wait
This method is primarily used by tests, and it was noticed that most of
the test cases complete in roughly 0.4s-0.9s in the normal case. Thus,
polling more frequently every 0.1s seems to strike a nice balance. With
this change, the tests seems to take on average 5s less time.