Improve allowed_storages filter spec
The previous spec did not excercise the fix because the created objects'
ems_metadata tree did not match the `belongsto_filter` so the
`MiqFilter.belongsto2object_list` returned an empty array which bypassed
the `MiqPreloader.preload_and_map(sources, :storages)` call which caused
the exception.

Now without the fix in EmsFolder to add the `virtual_has_many` this spec
will raise the same exception we observed during provisioning.