Add gitignores for rails credentials
We're switching to rails credentials to keep current with rails 7.1 and the future.
Note, we're assuming we won't want to commit and share encrypted credentials. If we want
to share them, such as for recording cassettes, the comments describe how to switch to only
ignoring the plain text encryption key files.

Followup to:  https://github.com/ManageIQ/manageiq/pull/23254
Required for: https://github.com/ManageIQ/manageiq-providers-autosde/pull/253