Fix metric/ci_mixin_spec using kubevirt vm
The metric/ci_mixin_spec was using a kubeivrt vm as an example of a
class which doesn't support capture.

This is no longer accurate as we have added C&U support for kubevirt,
plus we shouldn't be using provider plugin classes in core.

These specs are essentially testing that a class that includes a module
with `supports :capture` responds to `.supports?(:capture)` with true
which isn't actually testing anything.