Remove unnecessary test failing on rails 7.1.
Rails 7.1 detects this as a useless between clause, changing it
to a simple = operator with the timestamp value.

It was originally added in 6a94a1d93b801f88 #7075 and it seems
it was only to capture the current behavior at the time.  I think
it's testing internals of rails and not really important for testing
behavior of MiqExpression.

In rails 7.1, it was failing with:

       expected: "\"vms\".\"last_scan_on\" BETWEEN '2011-01-10 00:00:00' AND '2011-01-10 00:00:00'"
            got: "\"vms\".\"last_scan_on\" = '2011-01-10 00:00:00'"