Add interactive-ignore support through test:security:brakeman
Building the brakeman.ignore file requires us to have our patches in
place so that the values are built correctly to have the right
fingerprint. Thus, we can't just run brakeman -I directly. This change
allows us to pass `-I` or `--interactive-ignore` to the rake task, which
aids in updating the brakeman.ignore file.