Don't blow up on missing assets in production
In production, it's preferable for a missing asset to just gracefully
show a broken image than to blow out the call stack and potentially
break a page. This commit sets unknown_asset_fallback to true in
production mode to avoid blowing up.

In development and test, we leave it as the default, which will continue
to blow up, allowing us to find the missing assets.