Prevent orphans by purging associated records before the parent table
Alternative to #23381, where each purge batch was done in a transaction. Actually,
we only care about eventual consistency, so we can purge the associated records
first, and if that completes, we purge the parent table's rows for that batch.
This allows us to deal with timeouts or other errors by continuing a prior run
by cleaning up associated rows and only removing the parent when there are no more
associated rows in other tables that remain.

Note, there is no recourse for delete_all that has a failure as it's a vanilla DELETE
FROM where ... so it's very unlikely to fail for anything we care to handle.