Add purge_by_date_and_orphaned
Introduce purging first by date and then by orphanes.

It is not optimal to take 2 arguments, since most purging has stuck to one
But both of these need to be passed in.

We could have leveraged these 2 methods separately, but we wanted them to
be serialized, and avoid adding a dependency in the queue.