Opt into purging by destroy for container entities, use delete elsewhere
Fixes most of the issues in #23307

We were leaving around lots of orphaned container* rows when we removed
the container entities.  This change allows us to opt-into use destroy
on the primary table in situations where know the associated records
are NOT going to be many tens of thousands of rows.  If they are, those
associations should NOT be using dependent :destroy, and have their own
purger.