Purge request logs from requests, purge via destroy, drop batch size
Purge via destroy will ensure each request is destroyed without error.  If it fails,
such as from timeout, it will remain so it can be cleaned up the next run.

It should be safe to let request destroy to remove the request tasks and approvals
for the batch.

Row counts for example databases indicate requests, approvals and request tasks
seem to be roughly 1 to 1.

Request logs, on the other hand are many hundreds of request logs to 1 request.

Because many hundreds of logs can be found per request and requests are unlikedly to be
hundreds of thousands of rows, decrease the batch size to ensure each batch's list of ids
is as small as possible.