Quiet connections from the log
The connections checkout often output 5 lines for every connection. this puts a bunch into the logs.
When MIQ_QUIET is enabled, this will no longer be output to the logs

This is coded to work at class load time. It would need to be implemented differently if
you wanted to have this option in settings.