Fix issue where the code branch wasn't actually being tested
In rspec-mocks 3.13.3, a bug fix was made to correct the visibility of
stubbed methods. Since the visibility is now correct, the respond_to?
call no longer succeeds, and the stubbed method is not being called.
Since the stubbed method is now actually getting called, we need to fix
the test to actually work properly, such as fixing the order of the
parameters and attaching an actual cluster object.

This implies that this code actually never worked in non-test for this
specific event, because respond_to? could never return true on a
private method (although it was in test due to the method visibility
bug).