Add appliance/container tests for ansible-runner
This commit introduces Bats tests for ansible-runner to allow testing
the manageiq-ansible-venv installation. This tests cover ansible-runner
CLI tests as well as tests running through the Ansible::Runner Ruby
class, even without database access.

This commit also includes changes to Ansible::Runner itself to allow for
better detection of ansible and the manageiq-ansible-venv, allowing the
tests to run on the old appliances with python3.9 as well as the new
appliances with python3.12.