CVE-2025-27220 cgi gem
see also https://github.com/advisories/GHSA-mhwm-jh88-3gjf

Was a little concerned upgrading by so many versions.
You can see the diff https://github.com/ruby/cgi/compare/v0.1.0.2...v0.3.5.1
2020 to 2022 (Plus the cve fix in 2025)
Half the changes are to build environment.
Quite a few changes to regex
Added checks in case developer blindly put user data into cookies/headers

The version number is specific.
The suggestion was to upgrade to 0.3.5.1 and not 0.3.6
I did not verify that we wanted to avoid 0.3.6

NOTE: This vulnerability only affects Ruby 3.1 and 3.2, and we are running ruby < 3.0.
But the vulnerability detectors do not seem to take the ruby version into account
and pushing this change to avoid false positives