CVE-2024-43380 rufus-scheduler version
We need the latest version of fudit
Fudit is a dependency of rufus-scheduler.
This upgrade to rufus-scheduler to sets a minimum version for fudit and this resolves the CVE

It also sets a minimum version for et-orbit of 1.2.11, so we can remove our explicit requirement of 1.2.2

morphy is pegged to version 3.8.1. so it has the following changes:
https://github.com/jmettraux/rufus-scheduler/compare/v3.8.1...v3.9.2