includes uses tags while references does not
Over time, we have made more of a distinction of what goes into includes vs references.

This takes tags out of references.

This also adds tags to the invent_report_includes, which probably seems contradictory.
But this gives a more complete version of includes when it is not in the report. They will get filtered out
by include_as_hash