For inline_scopes, join to the references and exp_includes.
We wanted to use references, but includes is in a better form.
And due to the way we default references = includes
and rails just references all includes if references is anything (bug), it is good enough

Over the next few commits, we are being more deliberate about references, and we
are avoiding the whole includes().references() bug.

This is not tapping into any joins includes/joins that were part of the query before they
came into here. Which I wish were not the case. But in truth, we don't do that ever, so
supporting a non-used behavior that possibly is broken in other places is pointless