Reload after changing file paths to play nice with FactoryBotRails
FactoryBotRails calls find_definitions in it's railtie as rails initializes and
we can't assume we're the first one to call.  Instead we call reload, which
will reset things and then call find_definitions.

find_definitions isn't reentrant so if it's already run, a second call will try to
register the same factories from the same files and lead to the following
error:

```
 FactoryBot::DuplicateDefinitionError:
  Factory already registered: account
```