rbac inline_views detect collections (via a has_one) for distinct
For an inline view to work properly, the inner query needs to be properly distinct.
We had a test for this, but since the collection was a second tier, it didn't detect it

Before
======

Vm.includes(:hardware => :disks)
SELECT vms.* JOINS hardwares JOINS disks

We checked Vm.reflect_on_relation(:hardware).collection?
and since it is false, called it a day.

Luckily, for our test, we had :tags, where are a collection and it passed. (UGH)

After
=====

Now we traverse the whole hash (both `:hardware` and `:disks`)

 but since tags were passed in, it wasn't properly testing