Remove taggings and polymorphic from references
Now that we've moved form references.includes to preload and eager_load,
They are respecting the difference between those 2.

So we can pass polymorphic to preload and it won't blow up.
We can pass taggings to preload and it won't join and bring back too much data.

This commit updates our ar_references methods to filter out polymorphic and taggings from an includes hash.
Also removed virtual attributes since we can't access ruby virtual attributes in sql
(and sql virtual attributes are sub-selects so they shouldn't be in an includes anyway)

We can drop the PolymorphicError issues (handled in rbac filterer and report generator)

We could drop `includes_to_references` support but it is hanging around for legacy purposes and can be dropped soon.