Tests and comments around rbac and ActsAsArScope
Turns out for AAAS case, we currently only need to define all
Rbac has already used all, so all methods go directly to the aar_scope (which an an active record scope)
So all of these other methods are moot. Well, for this implementation.

We do want it to be complete in case the code changes and someone decides
to do something like call eager_load before calling all