Drop Log Collection Backend
In containerized deployments we can't support log collection, because we
can't mount the log depot. In appliance deployments, we're trying to get
away from the application UI having anything to do with the system it's
running on to avoid the need to elevate permissions. As such, this
commit removes log collection from backend. The UI will be done
separately.

Instead, on appliances, users can manually mount an NFS server (or
really any server type - even more than our UI supports), and then use
the /var/www/miq/vmdb/tools/collect_logs tool. If the user needs
scheduling, they can schedule a cron on the appliance to run the tool.